requirements.txt
anyio==3.5.0
appnope==0.1.3
argon2-cffi==21.3.0
asgiref==3.5.1
Babel==2.10.1
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==5.0.0
cachetools==5.0.0
calamus==0.4.0
certifi==2021.10.8
chardet==4.0.0
click==8.1.3
Cython==0.29.28
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
Django==3.2.4
django-cors-middleware==1.5.0
djangorestframework==3.13.1
djongo==1.3.6
entrypoints==0.4
fastjsonschema==2.15.3
frozendict==2.3.2
idna==2.10
importlib-metadata==4.11.3
ipykernel==6.13.0
ipython==8.3.0
ipython-genutils==0.2.0
ipywidgets==8.0.0rc0
isodate==0.6.1
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
json5==0.9.6
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.3.0
jupyter-console==6.4.3
jupyter-core==4.10.0
jupyter-server==1.17.0
jupyterlab==4.0.0a24
jupyterlab-pygments==0.2.2
jupyterlab-server==2.13.0
jupyterlab-widgets==3.0.0rc0
lazy-object-proxy==1.7.1
lxml==4.8.0
MarkupSafe==2.1.1
marshmallow==3.15.0
matplotlib-inline==0.1.3
mistune==0.8.4
nbclient==0.6.0
nbconvert==6.5.0
nbformat==5.3.0
nest-asyncio==1.5.5
nltk==3.5
notebook==7.0.0a3
notebook-shim==0.1.0
NRCLex==3.0.0
numpy==1.22.3
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
pipfreeze==2.0.3
prometheus-client==0.14.1
prompt-toolkit==3.0.29
psutil==5.9.0
Pygments==2.12.0
PyLD==2.0.3
pymongo==4.2.0
pyparsing==3.0.8
pyphen==0.12.0
python-dateutil==2.8.2
pytz==2022.1
pyzmq==23.0.0b1
qtconsole==5.3.0
QtPy==2.0.1
rdflib==6.1.1
regex==2022.4.24
requests==2.25.1
scikit-learn==0.24.2
scipy==1.8.0
Send2Trash==1.8.1b0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.2.4
stack-data==0.2.0
terminado==0.13.3
textblob==0.17.1
textstat==0.7.1
threadpoolctl==3.1.0
tinycss2==1.1.1
tornado==6.1
tqdm==4.64.0
traitlets==5.1.1
UNKNOWN==0.0.0
urllib3==1.26.9
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.2
widgetsnbextension==4.0.0rc0
